mobile/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
sqlalchemy==2.0.38
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
jinja2
edge-tts==3.1.5
